authentication

+++ server/users/routes/usersRestController.js
@@ -73,10 +73,14 @@ router.get("/:id", auth, async (req, res) => {
   }
 });
 
-router.patch("/:id", async (req, res) => {
+router.put("/:id", auth, async (req, res) => {
   try {
+    const { _id, isAdmin } = req.user;
     const userId = req.params.id;
     let user = req.body;
+
+    if (_id !== userId && !isAdmin)
+      throw new Error("only the user or admin can update the user details");
     user = normlizeUser(user);
     user = await updateUser(userId, user);
     return res.send(user);
@@ -85,9 +89,15 @@ router.patch("/:id", async (req, res) => {
   }
 });
 
-router.delete("/:id", async (req, res) => {
+router.delete("/:id", auth, async (req, res) => {
   try {
+    const { _id, isAdmin } = req.user;
     const userId = req.params.id;
+
+    if (_id !== userId && !isAdmin)
+      throw new Error(
+        "only the user who create the profile or an admin user can delete the user"
+      );
     const user = await deleteUser(userId);
     return res.send(user);
   } catch (error) {
